Major Change : Part 2
change the design for add implant page for more improved user experience. There are extra validation to guide user during the implant registration as well as during the generate icf. This changes are made for front-end only. The back end remain the same. Will continue on the update implant form front-end and the back end of all implant related. This includes the client review and comment to add the Approval type as dropdown menu. So the Approval Type model and structure also have been added. The Add and delete function also have been implemented. Currently the Approval Type features are implemented in Add implant and generate ICF only.